Move libs to deprecated named export
