+++ libs/rollup.config.js
@@ -30,7 +30,7 @@ const commonConfig = {
     'ethereumjs-wallet',
     'ethers/lib/utils',
     'ethers/lib/index',
-    'hashids/cjs',
+    'hashids/cjs'
   ]
 }
 
@@ -38,28 +38,29 @@ const commonTypeConfig = {
   plugins: [dts()]
 }
 
-export default [{
-  input: 'src/index.js',
-  output: [
-    { file: pkg.main, format: 'cjs', exports: 'auto', sourcemap: true }
-  ],
-  ...commonConfig
-},
-{
-  input: './src/types.ts',
-  output: [{ file: pkg.types, format: 'cjs' }],
-  ...commonTypeConfig
-},
-{
-  input: 'src/core.ts',
-  output: [
-    { file: pkg.core, format: 'cjs', exports: 'auto', sourcemap: true }
-  ],
-  ...commonConfig
-},
-{
-  input: './src/core.ts',
-  output: [{ file: pkg.coreTypes, format: 'cjs' }],
-  ...commonTypeConfig
-},
+export default [
+  {
+    input: 'src/index.js',
+    output: [
+      { file: pkg.main, format: 'cjs', exports: 'auto', sourcemap: true }
+    ],
+    ...commonConfig
+  },
+  {
+    input: './src/types.ts',
+    output: [{ file: pkg.types, format: 'cjs' }],
+    ...commonTypeConfig
+  },
+  {
+    input: 'src/core.ts',
+    output: [
+      { file: pkg.core, format: 'cjs', exports: 'auto', sourcemap: true }
+    ],
+    ...commonConfig
+  },
+  {
+    input: './src/core.ts',
+    output: [{ file: pkg.coreTypes, format: 'cjs' }],
+    ...commonTypeConfig
+  }
 ]
