+++ libs/src/libs.d.ts
@@ -1,0 +1,3 @@
+declare const libs: {}
+
+export default libs
