+++ libs/src/sdk/constants.ts
@@ -1,0 +1,11 @@
+export const CLAIM_DISTRIBUTION_CONTRACT_ADDRESS =
+  '0x683c19E621A0F107a291fdAB38f80179809d61B5'
+export const ETH_OWNER_WALLET = '0xC7310a03e930DD659E15305ed7e1F5Df0F0426C5'
+export const ETH_PROVIDER_URLS = 'https://eth.audius.co'
+export const ETH_REGISTRY_ADDRESS =
+  '0xd976d3b4f4e22a238c1A736b6612D22f17b6f64CREACT_APP_ETH_REGISTRY_ADDRESS=0xd976d3b4f4e22a238c1A736b6612D22f17b6f64C'
+export const ETH_TOKEN_ADDRESS = '0x18aAA7115705e8be94bfFEBDE57Af9BFc265B998'
+export const IDENTITY_SERVICE_ENDPOINT = 'https://identityservice.audius.co'
+export const REGISTRY_ADDRESS = '0xC611C82150b56E6e4Ec5973AcAbA8835Dd0d75A2'
+export const WEB3_PROVIDER_URL = 'https://poa-gateway.audius.co'
+export const WORMHOLE_ADDRESS = '0x6E7a1F7339bbB62b23D44797b63e4258d283E095'
