+++ libs/src/sdk/oauth.ts
@@ -1,0 +1,9 @@
+export const oauth = () => {
+  const loginWithAudius = () => {
+    // TODO
+    return null
+  }
+  return {
+    loginWithAudius
+  }
+}
