+++ libs/src/services/discoveryProvider/DiscoveryProvider.ts
@@ -29,6 +29,21 @@ type RequestParams = {
   data?: Record<string, unknown>
 }
 
+export type DiscoveryProviderConfig = {
+  whitelist?: Set<string>
+  blacklist?: Set<string>
+  userStateManager: UserStateManager
+  ethContracts: EthContracts
+  web3Manager?: Web3Manager
+  reselectTimeout?: number
+  selectionCallback?: DiscoveryProviderSelectionConfig['selectionCallback']
+  monitoringCallbacks?: DiscoveryProviderSelectionConfig['monitoringCallbacks']
+  selectionRequestTimeout?: number
+  selectionRequestRetries?: number
+  unhealthySlotDiffPlays?: number
+  unhealthyBlockDiff?: number
+}
+
 /**
  * Constructs a service class for a discovery node
  * @param whitelist whether or not to only include specified nodes in selection
@@ -50,34 +65,33 @@ export class DiscoveryProvider {
   blacklist: Set<string> | undefined
   userStateManager: UserStateManager
   ethContracts: EthContracts
-  web3Manager: Web3Manager
+  web3Manager: Web3Manager | undefined
   unhealthyBlockDiff: number
   serviceSelector: DiscoveryProviderSelection
-
   selectionRequestTimeout: number
   selectionRequestRetries: number
   unhealthySlotDiffPlays: number | undefined
   request404Count: number
   maxRequestsForTrue404: number
-  monitoringCallbacks: DiscoveryProviderSelection['monitoringCallbacks']
+  monitoringCallbacks:
+    | DiscoveryProviderSelection['monitoringCallbacks']
+    | undefined
   discoveryProviderEndpoint?: string
 
-  constructor(
-    whitelist: Set<string> | undefined,
-    blacklist: Set<string> | undefined,
-    userStateManager: UserStateManager,
-    ethContracts: EthContracts,
-    web3Manager: Web3Manager,
-    reselectTimeout: number | undefined,
-    selectionCallback:
-      | DiscoveryProviderSelectionConfig['selectionCallback']
-      | undefined,
-    monitoringCallbacks: DiscoveryProviderSelectionConfig['monitoringCallbacks'],
-    selectionRequestTimeout?: number,
-    selectionRequestRetries?: number,
-    unhealthySlotDiffPlays?: number,
-    unhealthyBlockDiff?: number
-  ) {
+  constructor({
+    whitelist,
+    blacklist,
+    userStateManager,
+    ethContracts,
+    web3Manager,
+    reselectTimeout,
+    selectionCallback,
+    monitoringCallbacks,
+    selectionRequestTimeout = REQUEST_TIMEOUT_MS,
+    selectionRequestRetries = MAX_MAKE_REQUEST_RETRY_COUNT,
+    unhealthySlotDiffPlays,
+    unhealthyBlockDiff
+  }: DiscoveryProviderConfig) {
     this.whitelist = whitelist
     this.blacklist = blacklist
     this.userStateManager = userStateManager
@@ -98,9 +112,8 @@ export class DiscoveryProvider {
       },
       this.ethContracts
     )
-    this.selectionRequestTimeout = selectionRequestTimeout ?? REQUEST_TIMEOUT_MS
-    this.selectionRequestRetries =
-      selectionRequestRetries ?? MAX_MAKE_REQUEST_RETRY_COUNT
+    this.selectionRequestTimeout = selectionRequestTimeout
+    this.selectionRequestRetries = selectionRequestRetries
     this.unhealthySlotDiffPlays = unhealthySlotDiffPlays
 
     // Keep track of the number of times a request 404s so we know when a true 404 occurs
@@ -110,7 +123,7 @@ export class DiscoveryProvider {
     this.request404Count = 0
     this.maxRequestsForTrue404 = MAX_MAKE_REQUEST_RETRIES_WITH_404
 
-    this.monitoringCallbacks = monitoringCallbacks ?? {}
+    this.monitoringCallbacks = monitoringCallbacks
   }
 
   async init() {
