+++ libs/src/services/identity/IdentityService.ts
@@ -58,12 +58,17 @@ type AttestationResult = {
 // Only probabilistically capture 50% of relay captchas
 const RELAY_CAPTCHA_SAMPLE_RATE = 0.5
 
+type IdentityServiceConfig = {
+  identityServiceEndpoint: string
+  captcha?: Captcha
+}
+
 export class IdentityService {
   identityServiceEndpoint: string
-  captcha: Captcha
+  captcha: Captcha | undefined
   web3Manager: Web3Manager | null
 
-  constructor(identityServiceEndpoint: string, captcha: Captcha) {
+  constructor({ identityServiceEndpoint, captcha }: IdentityServiceConfig) {
     this.identityServiceEndpoint = identityServiceEndpoint
     this.captcha = captcha
     this.web3Manager = null
