+++ libs/src/services/solanaAudiusData/index.ts
@@ -1171,5 +1171,3 @@ export class SolanaAudiusData {
     return await this.sendTx(tx)
   }
 }
-
-module.exports = SolanaAudiusData
